requirements.txt
django-multiupload
pandas
dj-database-url
black[d]
psycopg2-binary
pandera[io]
django-jsoneditor